embedding-model/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
numpy>=1.26,<3
optimum[onnxruntime]>=1.23,<2
transformers>=4.46,<5
python-dotenv>=1.0,<2
